% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate Entropy Statistics}
\usage{
entropy(tab)
}
\arguments{
\item{tab}{a numeric matrix}
}
\value{
a data frame with three columns: \code{eH}, \code{G}, and \code{E5}
}
\description{
Because it's possible to have multiple results with a minimum number of
samples, one way of assessing their importance is to calculate how
distributed the alleles are among the samples. This can be done with entropy
statistics
}
\examples{
data(monilinia)
entropy(monilinia)
set.seed(1999)
i <- find_samples(monilinia, n = 150, cut = TRUE)
entropy(monilinia[i[[1]], ])
entropy(monilinia[i[[2]], ])
}
