% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_samples.R
\name{find_samples}
\alias{find_samples}
\title{Iteratively find minimum set of samples by shuffling rows}
\usage{
find_samples(tab, n = 10, sort = TRUE, cut = FALSE)
}
\arguments{
\item{tab}{a numeric matrix with rownames}

\item{n}{the number of permutations to perform}

\item{sort}{when \code{TRUE} (default), the returned list will be sorted in order
of number of samples.}

\item{cut}{when \code{TRUE}, only the results with the minimum number of samples
will be returned.}
}
\value{
a list of character vectors
}
\description{
Because \code{\link[=get_minimum_set]{get_minimum_set()}} is deterministic, it may not present the
minimum set that represents all variables. This procedure automates the
process of randomly sampling the rows in the incoming matrix without
replacement to find a minimum set.
}
\examples{
data(monilinia)
# Iterate over the data 100 times and return only the minimum values
set.seed(2018)
find_samples(monilinia, n = 100, cut = TRUE)

# This is a random process and will not always return the same values
set.seed(201)
find_samples(monilinia, n = 100, cut = TRUE)
}
