% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repvar-package.r
\docType{package}
\name{repvar}
\alias{repvar}
\alias{repvar-package}
\title{repvar.}
\description{
This package allows you to find the minimum set of samples that represents
all non-zero variables in a binary matrix.
}
\details{
Minimum set of samples can be found with the function \code{\link[=rpv_find]{rpv_find()}}. This
function will shuffle your data set and pass it to \code{\link[=rpv_indices]{rpv_indices()}}. From
this, you will have a list of sample names.

Because there may be several combinations of samples that represent all
variables, the function \code{\link[=rpv_stats]{rpv_stats()}} can be used to calculate entropy
statistics over these variables.

If you want to visualize your data set, you can use \code{\link[=rpv_image]{rpv_image()}}
}
