% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minimum_set.R
\name{rpv_indices}
\alias{rpv_indices}
\title{Get minimum set of individual indices to represent all alleles in a population}
\usage{
rpv_indices(tab)
}
\arguments{
\item{tab}{an n x m matrix of individuals in rows and alleles in columns.}
}
\value{
a vector of integers representing row indices in the \code{tab}
}
\description{
Get minimum set of individual indices to represent all alleles in a population
}
\examples{
data(monilinia)
i <- rpv_indices(monilinia)
i
all(colSums(monilinia[i, ], na.rm = TRUE) > 0)
}
