% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repvar-package.r
\docType{package}
\name{repvar}
\alias{repvar}
\alias{repvar-package}
\title{repvar.}
\description{
This package allows you to find the minimum set of samples that represents
all non-zero variables in a binary matrix.
}
\details{
The underlying function is \code{\link[=get_minimum_set]{get_minimum_set()}}.
}
